Fixes and combine LR files
